% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterCDCData.R
\name{filter.CDCData}
\alias{filter.CDCData}
\title{This function takes the CDC data as input and filter based on the country input parameter}
\usage{
filter.CDCData(ds, country)
}
\arguments{
\item{ds}{CDC file}

\item{country}{Country}
}
\value{
streamlined file based on input country
}
\description{
Takes in CDC file, country as input paramaters and returns a filtered file for the country
}
