% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatCDCData.R
\name{format.CDCData}
\alias{format.CDCData}
\title{This function takes the filtered CDC data by inputed country as input and
will format the data as needed in order to generate a timeseries plot}
\usage{
\method{format}{CDCData}(cdcfilteredDs)
}
\arguments{
\item{cdcfilteredDs}{CDC file}
}
\value{
timeseries chart based on the file based on input country
}
\description{
Takes in filtered CDC file based on requested country and formats the data and returns
the data in the appropriate format to generate a plot
}
