% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCDCData.R
\name{getCDCData}
\alias{getCDCData}
\title{take input x (country), reads the world bank repository data for life expectancy at birth
and returns a timeseries chart for the inputed country for the period of 1960 to 2015}
\usage{
getCDCData(country)
}
\arguments{
\item{x}{input parameter containing the country requesting life timeseries}
}
\value{
timeseries chart for the inputed country for years 1960-2015
}
\description{
take input x (country), reads the world bank repository data for life expectancy at birth
and returns a timeseries chart for the inputed country for the period of 1960 to 2015
}
