% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveCDCData.R
\name{retrieve.CDCData}
\alias{retrieve.CDCData}
\title{This funtion will retrieve the Life expectancy at birth in total (years) CSV file from the CDC site}
\usage{
retrieve.CDCData()
}
\arguments{
\item{null}{not input parameter}
}
\value{
life expectancy file from CDC website
}
\description{
It does not take any parameters get the data and rename the columns
}
